Tests/LibWeb/Text/expected/grid/positioned-grid-descendants-left-static-top.txt

abspos1: 25,25 50x25
abspos2: 225,175 50x25
